Refactor hamiltonian abstractions
